add ~{} substitution for support of ldms config patterns
This improvement is compatible with existing json, yaml, and env
conventions. Hard-coded defaults which do not support all sampler use
cases are removed from the samplers conf. Samplers which require
ldms keywords (instead of attribute:value) options are now supported
by listing the keyword in the yaml as "'particular_keyword' : .keyword".
When converting to AVL kw & av list, the special value '.keyword'
causes the key/val pair to be treated as an ldms keyword parameter.
In addition, the format of sampler configuration is prettified so humans
can easily understand what is generated.

How the substitution works:
String values in dicts and lists containing ~{key} are expanded with
values of keys at the same or higher dict scope (narrowest dict scope wins
in event of tie). Combined with the yaml anchor syntax, this allows for very DRY
definition of samplers. The example of input is added in config/prod.yaml.