add 'maestro_ctrl' lable to the exception trap to avoid user confusion
maestro_ctrl prints a line number of its own code rather than the
line number of the offending yaml input. Labeling the "Line xxx"
message as a maestro_ctrl should reduce the time spent by users
trying to find that line number in the input yaml.

It doesn't look like there's an easy way to produce a line reference to the input.