`expand_hostlist()` with number support
`hostlist.py:expand_hostlist()` is not only used to expand hostlist, but
also other values such as ports. In the case of single port, the user
may enter the numeric value instead of string (411 instead of "411").
This patch modifies the function to also accept numbers (integers or
floats) for users' convenience.