Support for advertisers and producer listeners
Adds support for "advertisers", which are LDMSD's that advertise their connection
to known aggregators.
Adds support for producer listeners which are LDMSD aggregators that listen for an
advertised producer from which to aggregate data.
Add function to maestro_util that checks and returns arguments for authentication domains.
Fix bug in check_intrvl_str that reported an error for a valid time-interval string.